Allow executing external processes in task actions of build logic builds
Included plugin builds and buildSrc build may have task actions that
execute external processes. This is now allowed with configuration cache
enabled: if the included build is configuration-cache-compatible when
running standalone, it should not throw errors when running as part of
another build - even if technically external processes are spawned at
configuration time of the enclosing build.

This shifts the burden of ensuring correctnes to the build authors as
they have to properly declare inputs and outputs of tasks executing
external processes to make configuration-cached builds correct.