Allow executing external processes in build logic tasks preparation
The configuration cache already allows executing external processes at
upToDate, onlyIf and similar task callbacks for the main build. However,
build logic builds (buildSrc and included plugin builds) were only
allowed to execute external processes in task actions because of the way
the TaskExecutionTracker was listening to task execution.

This commit changes TaskExecutionTracker to listen to the build
operation that covers the whole execution of a task, including the
upToDate checks, thus making the behavior consistent.